+++ server/Whatta/src/main/java/whatta/Whatta/global/exception/ErrorCode.java
@@ -12,9 +12,10 @@ public enum ErrorCode {
     TIME_ORDER_INVALID(HttpStatus.BAD_REQUEST, "602", "끝나는 시간보다 시작 시간보다 앞서선 안됩니다."),
 
     USER_NOT_EXIST(HttpStatus.BAD_REQUEST, "800", "존재하지 않는 계정입니다."),
+    USER_SETTING_NOT_FOUND(HttpStatus.NOT_FOUND, "801", "해당 계정의 설정이 존재하지 않습니다."),
 
     LABEL_NOT_FOUND(HttpStatus.BAD_REQUEST, "900", "요청 라벨이 사용자의 라벨 목록에 없습니다."),
-    TOO_MANY_LABELS(HttpStatus.BAD_REQUEST, "900", "라벨은 최대 3개까지 설정할 수 있습니다.");
+    TOO_MANY_LABELS(HttpStatus.BAD_REQUEST, "901", "라벨은 최대 3개까지 설정할 수 있습니다.");
