+++ server/Whatta/src/main/java/whatta/Whatta/global/payload/Response.java
@@ -1,4 +1,4 @@
-package whatta.Whatta.global.payload.response;
+package whatta.Whatta.global.payload;
 
 import org.springframework.http.ResponseEntity;
