+++ server/Whatta/src/main/java/whatta/Whatta/global/repeat/Repeat.java
@@ -1,12 +1,9 @@
-package whatta.Whatta.global.entity;
+package whatta.Whatta.global.repeat;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import whatta.Whatta.global.anotation.ValidRepeat;
-import whatta.Whatta.global.enums.RepeatUnit;
-import whatta.Whatta.global.payload.request.RepeatRequest;
 
 import java.time.LocalDate;
 import java.util.List;
