+++ server/Whatta/src/main/java/whatta/Whatta/global/repeat/payload/RepeatRequest.java
@@ -1,10 +1,10 @@
-package whatta.Whatta.global.payload.request;
+package whatta.Whatta.global.repeat.payload;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import whatta.Whatta.global.anotation.ValidRepeat;
-import whatta.Whatta.global.entity.Repeat;
-import whatta.Whatta.global.enums.RepeatUnit;
+import whatta.Whatta.global.repeat.Repeat;
+import whatta.Whatta.global.repeat.RepeatUnit;
 
 import java.time.LocalDate;
 import java.util.List;
